PET-308 refactor : 배치 서비스 클래스 배치 모듈로 이동
